bg(MyDiary#Chall3): secure APIs with JWT and add documentation
 - created a bug branch for these changes
 - modified the db file by creating a foreign key in entries table from users table
 - modified the entry controllers file
 - modified the user controllers file
 - modified the authentication file
 - modified the validation file
 - modified the routes file
 - modified the test file
 - modified the readme file

 [ Finished-169339729 ]

+++ Server/Controllers/entryControllers.js
@@ -3,8 +3,8 @@ import { entryValidation } from '../Middlewares/valid';
 import { pool } from '../Config/db';
 
 const retrieveAllEntries = async (request, response) => {
-    const query = 'SELECT * FROM entries';
-    const view = await pool.query(query);
+    const query = 'SELECT * FROM entries WHERE uid = $1';
+    const view = await pool.query(query, [request.user.id]);
     return response.status(200).json({
         status: 200,
         message: 'Successful entries retrieve.',
@@ -14,8 +14,8 @@ const retrieveAllEntries = async (request, response) => {
 
 const retrieveSpecificEntry = async (request, response) => {
     const id = parseInt(request.params.id);
-    const values = [id];
-    const query = 'SELECT * FROM entries WHERE eid = $1';
+    const values = [id, request.user.id];
+    const query = 'SELECT * FROM entries WHERE eid = $1 AND uid = $2';
     const view = await pool.query(query, values);
     if (view.rows < '1') {
         response.status(404).json({
@@ -38,9 +38,10 @@ const addentry = async (request, response) => {
         return response.status(400).json({ status: 400, error: error.details[0].message });
     }
     const entryDate = moment().format('ll');
+    const author = request.user.id;
     const { entryTitle, posted, viewed, entryContent } = request.body;
-    const query = 'INSERT INTO entries(eTitle,eDate,ePosted,eViewed,eContent) VALUES($1,$2,$3,$4,$5) RETURNING *';
-    const values = [entryTitle, entryDate, posted, viewed, entryContent];
+    const query = 'INSERT INTO entries(eTitle,eDate,ePosted,eViewed,eContent,uid) VALUES($1,$2,$3,$4,$5,$6) RETURNING *';
+    const values = [entryTitle, entryDate, posted, viewed, entryContent,author];
     const add = await pool.query(query, values);
     return response.status(201).send({
         status: 201,
@@ -54,9 +55,10 @@ const modifyentry = async (request, response) => {
         if (error) {
             return response.status(400).json({ status: 400, error: error.details[0].message });
         }
+    const author = request.user.id;
     const { entryTitle, posted, viewed, entryContent } = request.body;
-    const query = 'UPDATE entries SET eTitle = $1, ePosted = $2, eViewed = $3, eContent = $4 WHERE eid = $5 RETURNING *';
-    const values = [entryTitle, posted, viewed, entryContent, parseInt(request.params.id)];
+    const query = 'UPDATE entries SET eTitle = $1, ePosted = $2, eViewed = $3, eContent = $4 WHERE eid = $5 AND uid = $6 RETURNING *';
+    const values = [entryTitle, posted, viewed, entryContent, parseInt(request.params.id), author];
     const edit = await pool.query(query, values);
     if (!edit.rows[0]) {
         response.status(404).json({
@@ -74,8 +76,8 @@ const modifyentry = async (request, response) => {
 };
 
 const entrydelete = async (request, response) => {
-    const query = 'DELETE FROM entries WHERE eid = $1 RETURNING *';
-    const value = [parseInt(request.params.id)];
+    const query = 'DELETE FROM entries WHERE eid = $1 AND uid = $2 RETURNING *';
+    const value = [parseInt(request.params.id), request.user.id];
     const del = await pool.query(query, value);
     if (!del.rows[0]) {
         response.status(404).json({
